+++ lib/modules/theme/theme_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: theme
-// Updated at: 2026-05-29 15:16:19
-// Commit: fix(theme): prevent double tap duplicate action in theme
+// Updated at: 2026-05-30 09:39:16
+// Commit: style(theme): harmonize light and dark theme palette for theme
 
 class themeComponentState {
-  final String commitRef = '1584';
-  final String lastMessage = 'fix(theme): prevent double tap duplicate action in theme';
-  final DateTime timestamp = DateTime.parse('2026-05-29T15:16:19+05:00');
+  final String commitRef = '1606';
+  final String lastMessage = 'style(theme): harmonize light and dark theme palette for theme';
+  final DateTime timestamp = DateTime.parse('2026-05-30T09:39:16+05:00');
 }